recommendation_system/requirements.txt
Django>=4.0
djangorestframework
djangorestframework-simplejwt
celery
redis
pandas
scikit-learn
